requirements.txt
aenum==3.1.15
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
audio-recorder-streamlit==0.0.10
Automat==24.8.1
backoff==2.2.1
bcrypt==4.2.0
blinker==1.8.2
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.17
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
constantly==23.10.4
cryptography==43.0.1
cssselect==1.2.0
dataclasses-json==0.6.7
deepgram-sdk==3.7.3
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
durationpy==0.9
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.2.2
faiss==1.8.0
Faker==30.1.0
fastapi==0.115.4
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.16.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
groq==0.11.0
grpcio==1.67.1
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.25.0
humanfriendly==10.0
hyperlink==21.0.0
idna==3.8
importlib_metadata==8.4.0
importlib_resources==6.4.5
incremental==24.7.2
itemadapter==0.9.0
itemloaders==1.3.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kubernetes==31.0.0
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.5
langchain-huggingface==0.1.0
langchain-text-splitters==0.3.0
langsmith==0.1.125
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.6.2
networkx==3.3
nltk==3.9.1
numpy @ file:///D:/bld/numpy_1707225570061/work/dist/numpy-1.26.4-cp310-cp310-win_amd64.whl#sha256=6761da75b1528684e6bf4dabdbdded9d1eb4d0e9b299482c7ce152cfb3155106
oauthlib==3.2.2
onnxruntime==1.19.2
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
overrides==7.7.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1718189413536/work
pandas==2.2.2
parsel==1.9.1
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
posthog==3.7.0
Protego==0.3.1
protobuf==4.25.5
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
PyDispatcher==2.0.7
pydub==0.25.1
Pygments==2.18.0
pyOpenSSL==24.2.1
pypdf==5.0.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
queuelib==1.7.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.8.0
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
Scrapy==2.11.2
sentence-transformers==3.1.1
service-identity==24.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.41.2
streamlit==1.38.0
streamlit-audiorec==0.1.3
streamlit-audiorecorder==0.0.5
streamlit-player==0.1.5
sympy==1.13.3
tenacity==8.5.0
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.2
toolz==1.0.0
torch==2.4.1
tornado==6.4.1
tqdm==4.66.5
transformers==4.44.2
Twisted==24.7.0
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.32.0
w3lib==2.2.1
watchdog==4.0.2
watchfiles==0.24.0
websocket-client==1.8.0
websockets==13.1
wrapt==1.16.0
yarl==1.11.1
zipp==3.20.2
zope.interface==7.1.1
